feat(briefings): move Add note + Briefing assistant to DesktopBottomBar
The desktop header had drifted from the Lovable design after the merge
with #1878 — it stacked three action buttons (Share + Add note +
Briefing assistant) on the right. Lovable keeps the desktop action
surface in a sticky bottom bar (mirroring the mobile bottom bar) and
leaves the header clean.

- DesktopBottomBar now owns the active-card-gated `openAddNoteTopLevel`
  + `openCardLevelChat` actions, with the same disabled state + title
  tooltips as the header used to have.
- DetailHeaderActions is reduced to the Share button; renders null when
  the share scope reports !canShare.
- Layout re-mounts DesktopBottomBar inside ShareScope; restores the
  outer `lg:pb-20` so content clears the bar.
